# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice SaaS API",
        "description": "API REST pour la plateforme de facturation multi-tenant Invoice SaaS. Authentification via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Invoice SaaS Support",
            "email": "support@invoice-saas.com"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://invoice-saas.com/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API principal"
        }
    ],
    "paths": {
...
        "/v1/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Initier un paiement pour une facture",
                "operationId": "c634c6fd47efbad2e7bc0511a70e514e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_id",
                                    "amount",
                                    "gateway",
                                    "currency",
                                    "return_url"
                                ],
                                "properties": {
                                    "invoice_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "enum": [
                                            "stripe",
                                            "paypal",
                                            "wave",
                                            "orange_money",
                                            "mtn_momo"
                                        ],
                                        "example": "orange_money"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "enum": [
                                            "XOF",
                                            "USD",
                                            "EUR"
                                        ],
                                        "example": "XOF"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "mobile_money",
                                        "nullable": true
                                    },
                                    "return_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://invoice-saas.com/payment/callback"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement initié — redirigez vers redirect_url",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment initiated successfully"
                                        },
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "https://pay.orange.ci/checkout/abc123"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 150000
                                                },
                                                "gateway": {
                                                    "type": "string",
                                                    "example": "orange_money"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur lors du traitement du paiement"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/payments/{id}/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirmer un paiement (callback depuis la passerelle)",
                "operationId": "05c8757d022cdf0156d6870808b0ce83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paiement",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Données de confirmation renvoyées par la passerelle",
                                "properties": {
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "txn_abc123"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement confirmé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment confirmed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 150000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "invoice": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Échec de la confirmation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}